collectief-ui/server/db/requirements.txt
paho-mqtt
uvicorn
fastapi
sqlalchemy
dotenv

